Add: 태그 realm db에 추가하는 기능 구현
