+++ TagIt/TagIt/ViewControllers/PageViewController.swift
@@ -25,6 +25,10 @@ class PageViewController: UIPageViewController {
             let viewControllers = [viewController]
             setViewControllers(viewControllers, direction: .forward, animated: false, completion: nil)
         }
+			
+			//test 값 임의로 추가하는 코드
+//			let photograph: Photograph = Photograph(name: "IMG_0033.JPG", localIdentifier: "69676321-F746-4F2C-9489-07522BA9EBC5/L0/001", colorId: "FDA293", tagArray: ["신발", "패션"])
+//			RealmManager.sharedInstance.saveObjects(object: photograph)
     }
     
     var targetSize: CGSize {
