+++ TagIt/TagIt/ViewControllers/ZoomedPhotoViewController.swift
@@ -59,7 +59,7 @@ class ZoomedPhotoViewController: UIViewController {
 					self.textView.text = tagString
 					self.textView.resolveHashTags()
 					self.textView.font = UIFont.systemFont(ofSize: 17.0)
-					self.textView.textColor = UIColor(hexFromString: photo.colorId)
+					self.textView.textColor = UIColor(hexFromString: PhotographManager.sharedInstance.selectedPhotograph!.colorId!)
 				}
 			}
     }
